Added logic for standalone and cluster recreate

+++ k8sutils/statefulset.go
@@ -39,6 +39,7 @@ type statefulSetParameters struct {
 	ExternalConfig        *string
 	ServiceAccountName    *string
 	UpdateStrategy        appsv1.StatefulSetUpdateStrategy
+	RecreateStatefulSet   bool
 }
 
 // containerParameters will define container input params
@@ -77,11 +78,11 @@ func CreateOrUpdateStateFul(namespace string, stsMeta metav1.ObjectMeta, params 
 		}
 		return err
 	}
-	return patchStatefulSet(storedStateful, statefulSetDef, namespace)
+	return patchStatefulSet(storedStateful, statefulSetDef, namespace, params.RecreateStatefulSet)
 }
 
 // patchStateFulSet will patch Redis Kubernetes StateFulSet
-func patchStatefulSet(storedStateful *appsv1.StatefulSet, newStateful *appsv1.StatefulSet, namespace string) error {
+func patchStatefulSet(storedStateful *appsv1.StatefulSet, newStateful *appsv1.StatefulSet, namespace string, recreateStateFulSet bool) error {
 	logger := statefulSetLogger(namespace, storedStateful.Name)
 
 	// We want to try and keep this atomic as possible.
@@ -178,7 +179,7 @@ func patchStatefulSet(storedStateful *appsv1.StatefulSet, newStateful *appsv1.St
 			logger.Error(err, "Unable to patch redis statefulset with comparison object")
 			return err
 		}
-		return updateStatefulSet(namespace, newStateful)
+		return updateStatefulSet(namespace, newStateful, recreateStateFulSet)
 	}
 	logger.Info("Reconciliation Complete, no Changes required.")
 	return nil
@@ -522,19 +523,21 @@ func createStatefulSet(namespace string, stateful *appsv1.StatefulSet) error {
 }
 
 // updateStatefulSet is a method to update statefulset in Kubernetes
-func updateStatefulSet(namespace string, stateful *appsv1.StatefulSet) error {
+func updateStatefulSet(namespace string, stateful *appsv1.StatefulSet, recreateStateFulSet bool) error {
 	logger := statefulSetLogger(namespace, stateful.Name)
 	_, err := generateK8sClient().AppsV1().StatefulSets(namespace).Update(context.TODO(), stateful, metav1.UpdateOptions{})
-	sErr, ok := err.(*apierrors.StatusError)
-	if ok && sErr.ErrStatus.Code == 422 && sErr.ErrStatus.Reason == metav1.StatusReasonInvalid {
-		failMsg := make([]string, len(sErr.ErrStatus.Details.Causes))
-		for messageCount, cause := range sErr.ErrStatus.Details.Causes {
-			failMsg[messageCount] = cause.Message
-		}
-		logger.Info("recreating StatefulSet because the update operation wasn't possible", "reason", strings.Join(failMsg, ", "))
-		propagationPolicy := metav1.DeletePropagationForeground
-		if err := generateK8sClient().AppsV1().StatefulSets(namespace).Delete(context.TODO(), stateful.GetName(), metav1.DeleteOptions{PropagationPolicy: &propagationPolicy}); err != nil {
-			return errors.Wrap(err, "failed to delete StatefulSet to avoid forbidden action")
+	if recreateStateFulSet {
+		sErr, ok := err.(*apierrors.StatusError)
+		if ok && sErr.ErrStatus.Code == 422 && sErr.ErrStatus.Reason == metav1.StatusReasonInvalid {
+			failMsg := make([]string, len(sErr.ErrStatus.Details.Causes))
+			for messageCount, cause := range sErr.ErrStatus.Details.Causes {
+				failMsg[messageCount] = cause.Message
+			}
+			logger.Info("recreating StatefulSet because the update operation wasn't possible", "reason", strings.Join(failMsg, ", "))
+			propagationPolicy := metav1.DeletePropagationForeground
+			if err := generateK8sClient().AppsV1().StatefulSets(namespace).Delete(context.TODO(), stateful.GetName(), metav1.DeleteOptions{PropagationPolicy: &propagationPolicy}); err != nil {
+				return errors.Wrap(err, "failed to delete StatefulSet to avoid forbidden action")
+			}
 		}
 	}
 	if err != nil {
